# Address recent performance regression